Address CodeRabbit review feedback
Correctness:
- calculate_transformation_errors: features that yielded no usable link now get
  null residuals instead of 0, which read as perfect control points; the pivot
  tolerance is scaled by matrix magnitude so degeneracy is detected in UTM
  metres and not false-positived on sub-metre grids.
- optimal_corridor_connections: per-axis step length (min(cell_x, cell_y) for
  both axes under-costed every move along the longer axis); Prim now grows a
  forest over all components, so an isolated region 0 no longer wipes out the
  whole network; NaN/inf rejected in validation; derived grid capped.
- spatially_constrained_multivariate_clustering: SNAP 1e-9 -> 1e-6 (1e-9 made
  contiguity exact-match, silently splitting the graph) with an i64 saturation
  guard; missing/non-numeric analysis values are now counted and reported.
- construct_sight_lines: MAX_SIGHT_LINES was bypassed on the join_field path.
- dissolve_route_events: dissolve mode declared the source field type but wrote
  Text; both modes now declare Text. route_count was the group count, so
  group_count is reported separately and route_count counts distinct routes.
- pivot_table / split_by_features: name disambiguation registered only the base,
  so ["a","a","a_1"] emitted a_1 twice — a duplicate schema field and, for
  split_by_features, a silently overwritten output file. Pivot columns are also
  now deconflicted against the identity fields.
- generate_points_along_3d_lines: generated LINE_ID/POINT_SEQ/CHAINAGE could
  collide with carried-through input fields (common on chainage data).
- subset_features: modulo now in u64; `as usize` first truncated on wasm32,
  breaking the documented cross-platform reproducibility.
- summarize_percent_change: corpus totals come from the deduplicated point
  lists, so overlapping catchments no longer double count.
- build_seamlines: band index validated against every raster up front.
- minimum_bounding_volume: documented cap on per-group hull input.
- idw_3d: validation predicate and message now agree.
- README: escaped literal pipes that broke the table row (MD056).
- lib.rs: split_by_features.output_dir typed file_out(), matching the other
  directory-output tools.

Performance:
- generate_breach_lines: full-raster buffers hoisted out of the per-depression
  loop and reset only where touched; connection points in a HashSet.
- adjust_stream_to_raster: the least-cost fallback is confined to a window
  around the node pair instead of sweeping and allocating the whole raster.
- spatially_constrained_multivariate_clustering: Prim uses a BinaryHeap; the
  cut search reuses a per-cluster adjacency instead of rebuilding one per
  candidate edge; vertex dedupe uses a HashSet.
- merge_divided_roads: merge keys precomputed, bbox pruning before the sampling
  test, HashSet vertex dedupe; displacement layer no longer declares Polygon
  when a MultiPolygon is possible.
- optimal_corridor_connections: single pass over cells for all region pairs;
  balanced pairwise union in thicken(), joint circles only where the direction
  changes; centerlines always returned.
- split_by_features: the zone geometry is converted once per zone rather than
  once per tested point.

Tests: two new optimal_corridor_connections cases, both verified to fail
against the pre-fix code — anisotropic cells (ACCUM_COST 52 vs true length 260)
and an isolated first region. Plus a summarize_percent_change case pinning the
overlapping-catchment semantics. 1448 tests pass; new modules clippy-clean.
Re-validated on the real data: SKATER, corridor, adjust_stream and split all
reproduce their earlier results, and generate_breach_lines is bit-identical.

+++ crates/geolibre-tools/src/subset_features.rs
@@ -160,7 +160,11 @@ impl Tool for SubsetFeaturesTool {
                 }
             };
             for k in 0..take {
-                let j = k + (rng.next_u64() as usize) % (bn - k);
+                // Modulo in u64: `as usize` first would keep only the low 32
+                // bits on wasm32, so the same seed would pick different indices
+                // in the browser than natively — exactly what this tool's
+                // reproducibility contract rules out.
+                let j = k + (rng.next_u64() % (bn - k) as u64) as usize;
                 bucket.swap(k, j);
                 is_training[bucket[k]] = true;
             }
